Relocation of FTP classes
