+++ boardgamesManagement/src/main/java/com/albertsalud/bgmanagement/utils/ftp/FTPServices.java
@@ -1,4 +1,4 @@
-package com.albertsalud.bgmanagement.utils;
+package com.albertsalud.bgmanagement.utils.ftp;
 
 import java.io.File;
 import java.io.IOException;
